Se termina endpoint que devuelve los datos de resumen del programa

+++ database/migrations/1588830713-create-base-parts.js
@@ -22,14 +22,14 @@ module.exports = {
 				onUpdate: 'cascade',
 				onDelete: 'cascade'
 			},
-			planned_lines_base: { type: Sequelize.DOUBLE, allowNull: false },
-			planned_lines_deleted: { type: Sequelize.DOUBLE, allowNull: false },
-			planned_lines_edits: { type: Sequelize.DOUBLE, allowNull: false },
-			planned_lines_added: { type: Sequelize.DOUBLE, allowNull: false },
-			current_lines_base: { type: Sequelize.DOUBLE },
-			current_lines_deleted: { type: Sequelize.DOUBLE },
-			current_lines_edits: { type: Sequelize.DOUBLE },
-			current_lines_added: { type: Sequelize.DOUBLE },
+			planned_lines_base: { type: Sequelize.FLOAT, allowNull: false },
+			planned_lines_deleted: { type: Sequelize.FLOAT, allowNull: false },
+			planned_lines_edits: { type: Sequelize.FLOAT, allowNull: false },
+			planned_lines_added: { type: Sequelize.FLOAT, allowNull: false },
+			current_lines_base: { type: Sequelize.FLOAT },
+			current_lines_deleted: { type: Sequelize.FLOAT },
+			current_lines_edits: { type: Sequelize.FLOAT },
+			current_lines_added: { type: Sequelize.FLOAT },
 			created_at: { allowNull: false, type: Sequelize.DATE },
 			updated_at: { allowNull: false, type: Sequelize.DATE }
 		})
